# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "工业机器人课程智能助教",
    "description": "面向工业机器人教学的可评测、可观测、受控 Agent 平台。",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "summary": "Create Chat",
        "operationId": "create_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/stream": {
      "get": {
        "summary": "Stream Run",
        "operationId": "stream_run_api_v1_runs__run_id__stream_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "User Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/feedback": {
      "post": {
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_api_v1_runs__run_id__feedback_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents": {
      "post": {
        "summary": "Import Document",
        "operationId": "import_document_api_v1_knowledge_documents_post",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentImportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_knowledge_documents_get",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/traces/{request_id}": {
      "get": {
        "summary": "Get Trace",
        "operationId": "get_trace_api_v1_traces__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bad-cases": {
      "get": {
        "summary": "List Bad Cases",
        "operationId": "list_bad_cases_api_v1_bad_cases_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "role": {
            "type": "string",
            "pattern": "^(student|teacher|maintainer)$",
            "title": "Role",
            "default": "student"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunAccepted": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "run_id",
          "status",
          "stream_url"
        ],
        "title": "RunAccepted"
      },
...
    }
  }
}